+++ rsProgramFragment.h
@@ -29,7 +29,7 @@ class ProgramFragment : public Program {
 public:
     ProgramFragment(Context *rsc, const char * shaderText, size_t shaderLength,
                     const char** textureNames, size_t textureNamesCount, const size_t *textureNamesLength,
-                     const uint32_t * params, size_t paramLength);
+                    const uintptr_t * params, size_t paramLength);
     virtual ~ProgramFragment();
 
     virtual void setup(Context *, ProgramFragmentState *);
